ScienceClaw/backend/requirements.txt
fastapi==0.128.7
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
sse-starlette==3.2.0
loguru==0.7.3
python-json-logger==4.0.0
bcrypt==5.0.0
shortuuid==1.0.13
motor==3.7.1
langchain
langchain-core
langchain-community==0.4.1
langchain-openai==1.1.8
langgraph==1.0.8
httpx==0.28.1
tavily-python==0.7.21
pyyaml>=6.0
deepagents==0.4.4
langchain-mcp-adapters
tooluniverse
lark-oapi
